A lot of improvements
- Fully recoded the players database when joining
- Added argument for givesaveditems to delete from file also
- Added new argument: /rm listplayers to list the in-game players
- Fixed value when set the points for player and the amount is 0 then added for player.
- Some fixes in the reload command to reload the plugin successfully.
- Some fixes in actionbar & bossbar if the value is not a boolean
- Some minor fixes in the MySQL
- Killed players with tool broadcast message should work now
- Fixed issue when players in-lobby then got damage from others
- 1.14 fix for items are dropped from inventory
- Fixed issue when the player still is in respawn screen then remove player if the game end.
- Decreased the egg explosion rate
- Now prevents the usage of bush
- Lobby timer should cancel at game start
- More actionbar messages
- Random teleport fixes, when does not teleported the players randomly
- Added option to send broadcast messages when the game end
- Should fixed scoreboard duplicated lines on player death
- Lobby timer will cancel automatically if the players size less than 1
- Actionbar should work
- Some fixes related to joining a player to a game
- Fixed point losing amounts
- Fixed issue when the signs are not removed correctly
- Players database now correctly loaded at startup

+++ src/main/java/hu/montlikadani/RageMode/gameUtils/BossMessenger.java
@@ -39,7 +39,7 @@ public class BossMessenger {
 					else
 						boss.removeAll();
 				}
-			}, (long) (20 * i));
+			}, 20 * i);
 		}
 	}
 }
